Node chooser returns a list of nodes
This ended up being a much larger refactoring than I initially anticipated. The overall goal is simple though: instead of randomly choosing nodes, potentially rechoosing the same nodes multiple times for a single query, get a full list of nodes and try them sequentially.

This ended up having a bunch of add-on impacts in the code, such as identifying that the simulation and broadcast sequence tracking information should be Cosmos-wide instead of at the Node level.

Additionally, the logic around all-nodes broadcast is slightly better now. If any of the broadcasts succeed, we will use it, even if broadcasting on selected nodes failed.