Revamped broadcast error checking
The issue here is that, during broadcast, we want to broadcast to multiple nodes, only returning an error if they _all_ fail. The way we did that previously was essentially:

* Spawn a bunch of tasks
* Each task puts their raw results on a channel
* A low-level functions looks for the first task result that is Ok (meaning: successfully performed the query)
* Perform additional error checks on the returned status code

The problem we were facing was that some nodes were very quickly failing on the additional error checks, but since they were working at the network layer, we treated those as "successes."

Instead now, when popping results off the channel, we perform the deeper status code checks immediately. That way, if any of the nodes succeed, the broadcast overall succeeds.