Allow for configuration of code IDs
This moves the handling of configured code IDs into the core library instead of the executable. Instead of simply hard-coding them, it uses the same configuration system as other values. This allows us to provide code IDs for custom networks or override the defaults.

This also adds CLI options for providing code IDs when generating a new CW3 flex contract.